+++ src/r4300/cop1_s.c
@@ -100,7 +100,7 @@ void ROUND_L_S(void)
 {
    if (check_cop1_unusable()) return;
    set_round();
-   *((long long*)(reg_cop1_double[cffd])) = *reg_cop1_simple[cffs];
+   *((long long*)(reg_cop1_double[cffd])) = (long long) *reg_cop1_simple[cffs];
    PC++;
 }
 
@@ -108,7 +108,7 @@ void TRUNC_L_S(void)
 {
    if (check_cop1_unusable()) return;
    set_trunc();
-   *((long long*)(reg_cop1_double[cffd])) = *reg_cop1_simple[cffs];
+   *((long long*)(reg_cop1_double[cffd])) = (long long) *reg_cop1_simple[cffs];
    PC++;
 }
 
@@ -116,7 +116,7 @@ void CEIL_L_S(void)
 {
    if (check_cop1_unusable()) return;
    set_ceil();
-   *((long long*)(reg_cop1_double[cffd])) = *reg_cop1_simple[cffs];
+   *((long long*)(reg_cop1_double[cffd])) = (long long) *reg_cop1_simple[cffs];
    PC++;
 }
 
@@ -124,7 +124,7 @@ void FLOOR_L_S(void)
 {
    if (check_cop1_unusable()) return;
    set_floor();
-   *((long long*)(reg_cop1_double[cffd])) = *reg_cop1_simple[cffs];
+   *((long long*)(reg_cop1_double[cffd])) = (long long) *reg_cop1_simple[cffs];
    PC++;
 }
 
@@ -132,7 +132,7 @@ void ROUND_W_S(void)
 {
    if (check_cop1_unusable()) return;
    set_round();
-   *((int*)reg_cop1_simple[cffd]) = *reg_cop1_simple[cffs];
+   *((int*)reg_cop1_simple[cffd]) = (int) *reg_cop1_simple[cffs];
    PC++;
 }
 
@@ -140,7 +140,7 @@ void TRUNC_W_S(void)
 {
    if (check_cop1_unusable()) return;
    set_trunc();
-   *((int*)reg_cop1_simple[cffd]) = *reg_cop1_simple[cffs];
+   *((int*)reg_cop1_simple[cffd]) = (int) *reg_cop1_simple[cffs];
    PC++;
 }
 
@@ -148,7 +148,7 @@ void CEIL_W_S(void)
 {
    if (check_cop1_unusable()) return;
    set_ceil();
-   *((int*)reg_cop1_simple[cffd]) = *reg_cop1_simple[cffs];
+   *((int*)reg_cop1_simple[cffd]) = (int) *reg_cop1_simple[cffs];
    PC++;
 }
 
@@ -156,7 +156,7 @@ void FLOOR_W_S(void)
 {
    if (check_cop1_unusable()) return;
    set_floor();
-   *((int*)reg_cop1_simple[cffd]) = *reg_cop1_simple[cffs];
+   *((int*)reg_cop1_simple[cffd]) = (int) *reg_cop1_simple[cffs];
    PC++;
 }
 
@@ -172,7 +172,7 @@ void CVT_W_S(void)
 {
    if (check_cop1_unusable()) return;
    set_rounding();
-   *((int*)reg_cop1_simple[cffd]) = *reg_cop1_simple[cffs];
+   *((int*)reg_cop1_simple[cffd]) = (int) *reg_cop1_simple[cffs];
    PC++;
 }
 
@@ -180,7 +180,7 @@ void CVT_L_S(void)
 {
    if (check_cop1_unusable()) return;
    set_rounding();
-   *((long long*)(reg_cop1_double[cffd])) = *reg_cop1_simple[cffs];
+   *((long long*)(reg_cop1_double[cffd])) = (long long) *reg_cop1_simple[cffs];
    PC++;
 }
